+++ wallet-new/src/Cardano/Wallet/API/V1/Types.hs
@@ -92,14 +92,17 @@ import           Universum
 
 import           Data.Semigroup (Semigroup)
 
+import           Cardano.Wallet.API.V1.Swagger.Example (Example, example)
 import           Control.Lens (At, Index, IxValue, at, ix, makePrisms, to, (?~))
 import           Data.Aeson
 import qualified Data.Aeson.Options as Serokell
 import           Data.Aeson.TH as A
 import           Data.Aeson.Types (toJSONKeyText, typeMismatch)
 import qualified Data.Char as C
+import           Data.Default (Default (def))
 import qualified Data.IxSet.Typed as IxSet
-import           Data.Swagger as S
+import           Data.Swagger hiding (Example, example)
+import qualified Data.Swagger as S
 import           Data.Swagger.Declare (Declare, look)
 import           Data.Swagger.Internal.Schema (GToSchema)
 import           Data.Swagger.Internal.TypeShape (GenericHasSimpleShape,
@@ -179,7 +182,7 @@ type IsPropertiesMap m =
 
 genericSchemaDroppingPrefix
     :: forall a m proxy.
-    ( Generic a, ToJSON a, Arbitrary a, GToSchema (Rep a), IsPropertiesMap m
+    ( Generic a, ToJSON a, Example a, GToSchema (Rep a), IsPropertiesMap m
     , GenericHasSimpleShape
         a
         "genericDeclareNamedSchemaUnrestricted"
@@ -196,10 +199,10 @@ genericSchemaDroppingPrefix prfx extraDoc proxy = do
     defs <- look
     pure $ s
       & over schema (over properties (extraDoc (addFieldDescription defs)))
-      & schema . example ?~ toJSON (genExample :: a)
+      & schema . S.example ?~ toJSON (genExample :: a)
   where
     genExample =
-      (unGen (resize 3 arbitrary)) (mkQCGen 42) 42
+      (unGen (resize 3 example)) (mkQCGen 42) 42
 
     addFieldDescription defs field desc =
       over (at field) (addDescription defs field desc)
@@ -710,6 +713,7 @@ deriveSafeBuildable ''SyncProgress
 instance BuildableSafeGen SyncProgress where
     buildSafeGen _ sp = bprint build sp
 
+instance Example SyncProgress where
 instance Arbitrary SyncProgress where
   arbitrary = SyncProgress <$> arbitrary
                            <*> arbitrary
@@ -1944,3 +1948,116 @@ instance Arbitrary Redemption where
 --
 
 type family Update (original :: *) :: * where
+  Update Wallet        = WalletUpdate
+  Update Account       = AccountUpdate
+  Update WalletAddress = () -- read-only
+
+type family New (original :: *) :: * where
+  New Wallet  = NewWallet
+  New Account = NewAccount
+  New WalletAddress = NewAddress
+
+type CaptureWalletId = Capture "walletId" WalletId
+
+type CaptureAccountId = Capture "accountId" AccountIndex
+
+
+--
+-- Example typeclass instances
+--
+
+instance Example (V1 Core.PassPhrase)
+instance Example (V1 Core.Coin)
+
+instance Example a => Example (WalletResponse a) where
+    example = WalletResponse <$> example
+                             <*> pure SuccessStatus
+                             <*> example
+
+-- | We have a specific 'Example' instance for @'V1' 'Address'@ because we want
+-- to control the length of the examples. It is possible for the encoded length
+-- to become huge, up to 1000+ bytes, if the 'UnsafeMultiKeyDistr' constructor
+-- is used. We do not use this constructor, which keeps the address between
+-- ~80-150 bytes long.
+instance Example (V1 Address) where
+    example = fmap V1 . Core.makeAddress
+        <$> arbitrary
+        <*> arbitraryAttributes
+      where
+        arbitraryAttributes =
+            Core.AddrAttributes
+                <$> arbitrary
+                <*> oneof
+                    [ pure Core.BootstrapEraDistr
+                    , Core.SingleKeyDistr <$> arbitrary
+                    ]
+
+instance Example BackupPhrase where
+    example = pure (BackupPhrase def)
+
+instance Example Address
+instance Example Metadata
+instance Example AccountIndex
+instance Example AccountBalance
+instance Example AccountAddresses
+instance Example WalletId
+instance Example AssuranceLevel
+instance Example SyncPercentage
+instance Example BlockchainHeight
+instance Example LocalTimeDifference
+instance Example PaymentDistribution
+instance Example AccountUpdate
+instance Example Wallet
+instance Example WalletUpdate
+instance Example WalletOperation
+instance Example PasswordUpdate
+instance Example EstimatedFees
+instance Example Transaction
+instance Example WalletSoftwareUpdate
+instance Example NodeSettings
+instance Example SlotDuration
+instance Example WalletAddress
+instance Example NewAccount
+instance Example TimeInfo
+instance Example AddressValidity
+instance Example NewAddress
+instance Example CUpdateInfo
+instance Example SubscriptionStatus
+instance Example NodeId
+
+instance Example InputSelectionPolicy where
+    example = pure OptimizeForHighThroughput
+
+instance Example (V1 InputSelectionPolicy) where
+    example = pure (V1 OptimizeForHighThroughput)
+
+instance Example Account where
+    example = Account <$> example
+                      <*> example -- NOTE: this will produce non empty list
+                      <*> example
+                      <*> pure "My account"
+                      <*> example
+
+instance Example NewWallet where
+    example = NewWallet <$> example
+                        <*> example -- Note: will produce `Just a`
+                        <*> example
+                        <*> pure "My Wallet"
+                        <*> example
+
+instance Example NodeInfo where
+    example = NodeInfo <$> example
+                       <*> example  -- NOTE: will produce `Just a`
+                       <*> example
+                       <*> example
+                       <*> example
+
+instance Example PaymentSource where
+    example = PaymentSource <$> example
+                            <*> example
+
+instance Example Payment where
+    example = Payment <$> example
+                      <*> example
+                      <*> example -- TODO: will produce `Just groupingPolicy`
+                      <*> example
